requirements.txt
Bottleneck==1.3.5
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
matlabengineforpython===R2022a
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scipy==1.9.0
six==1.16.0
toml==0.10.2
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.11
